feat: add db start command for ci
